% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tienmt.R
\name{tobit_bayes}
\alias{tobit_bayes}
\title{Gibbs Sampler for Tobit Regression with Horseshoe Prior}
\usage{
tobit_bayes(
  y,
  X,
  c_sensored = 0,
  n_iter = 2000,
  burn_in = 500,
  a0 = 1,
  b0 = 0.1,
  level = 0.95
)
}
\arguments{
\item{y}{Response vector (numeric) of length n, possibly censored.}

\item{X}{Design matrix (numeric), n x p.}

\item{c_sensored}{Censoring threshold (default 0).}

\item{n_iter}{Number of MCMC iterations (default 2000).}

\item{burn_in}{Number of burn-in iterations (default 500).}

\item{a0}{Shape parameter for inverse-gamma prior on sigma² (default 1).}

\item{b0}{Rate parameter for inverse-gamma prior on sigma² (default 0.1).}

\item{level}{level for return credible intervals and variable selction, default is 0.95. should be in (0,1).}
}
\value{
A list containing:
\describe{
  \item{beta_samples}{A numeric matrix of dimension \code{(n_iter - burn_in) × p} containing posterior samples of the regression coefficients after burn-in.}
  \item{full_beta}{A numeric matrix of dimension \code{n_iter × p} storing all sampled beta values including burn-in iterations. Useful for convergence diagnostics.}
  \item{sigma2HS}{A numeric value giving the posterior mean of the error variance \eqn{\sigma^2} over the samples after burn-in.}
  \item{selected}{A binary vector giving the selected variable based on credible intervals.}
}
}
\description{
Gibbs Sampler for Tobit Regression with Horseshoe Prior
}
\examples{
\dontrun{
  set.seed(1)
  X <- matrix(rnorm(100*5), 100, 5)
  beta0 <- c(2, -1, 0, 0, 0)
  y <- X \%*\% beta0 + rnorm(100)
  y[y < 0] <- 0
  res <- tobit_bayes(y, X)
  (posterior_means <- colMeans(res$beta_samples))
  plot(posterior_means, type = 'h', main = "Posterior Means of Beta", ylab = "Mean")
  res$selected
}
}
